Add support for trafficlight4
The latest version of the USB-TrafficLight4 has a different product
ID and allows setting all the directions of the traffic light
independently.

The new traffic light supports the old interface, but acts like the
previous model by shinning light in all directions at the same time.
To be able to select the directions, we need to use another interface:
the switch one which exposes all LEDs through a 16 bits bitfield.

This commit introduces support for the trafficlight4, using the new
switch interface... but only emulating the previous interface. This
is because the next commit will add support for selecting the
direction(s) we want to shine light to.

Signed-off-by: Martin Roukala (né Peres) <martin.roukala@mupuf.org>