Add support for directionality
In the previous commit, we added support for the trafficlight4 which
supports directionality but we could not make use for in the current
codebase.

In this commit, we add a new directionality parameter to the set_leds()
method. It is defaulting to the "ALL" directions, and is added at the
end of the argument list for maximum backwards compatibility with the
older interface.

The CLI tool has got modified to expose this feature using the
-d/--direction argument.

Signed-off-by: Martin Roukala (né Peres) <martin.roukala@mupuf.org>